+++ packages/core/src/agent/certificate.test.ts
@@ -1155,3 +1155,8 @@ describe('subnet certificates', () => {
     }
   });
 });
+
+test('hashTreeToString', async () => {
+  const cert = cbor.decode<Cert.Cert>(SAMPLE_CERT_BYTES);
+  expect(typeof Cert.hashTreeToString(cert.tree)).toBe('string');
+});